Decompose SurveyScreen.test to different files
